requirements.txt
selenium>=2.26.0
zmqdecorators>=0.7.1
pyzmq>=14.0
pyaml>=14.05.7
